requirements.txt
streamlit
pandas
plotly
scikit-learn
joblib
fastapi
uvicorn
pydantic
google-generativeai
requests